Moved temp init into GridProperty
The GridPropertyInitializers files now really only held the temperature
initialisation function, and has been merged into GridProperty.
